+++ efficientdetection/DetectK4c.java
@@ -92,10 +92,10 @@ public class DetectK4c {
 			
 			//make graph from new list of vertices
 			UndirectedGraph graph2 = Utility.makeGraphFromVertexSet(graph, nXList);
-			double [][] adjMatrix = graph2.getAdjacencyMatrix();
+			int[][] adjMatrix = graph2.getAdjacencyMatrix();
 			
 			//get square of adjacency matrix
-			double[][] adjMatrixSquare;
+			int[][] adjMatrixSquare;
 			try {
 				adjMatrixSquare = Utility.multiplyMatrix(adjMatrix, adjMatrix);
 			} catch (MatrixException e) {
@@ -160,10 +160,10 @@ public class DetectK4c {
 				continue;
 			
 			UndirectedGraph<Integer,Integer> graph2 = Utility.makeGraphFromVertexSet(graph, nXList);
-			double [][] adjMatrix = graph2.getAdjacencyMatrix();
+			int[][] adjMatrix = graph2.getAdjacencyMatrix();
 			
 			//get square of adjacency matrix
-			double[][] adjMatrixSquare;
+			int[][] adjMatrixSquare;
 			try {
 				adjMatrixSquare = Utility.multiplyMatrix(adjMatrix, adjMatrix);
 			} catch (MatrixException e) {
